Adds fzf support for list input
The list input can be quite challenging in some cases.
For instance, try to to override a method from one of your test case.
Given the number of method in PHPUnit TestCase class, vim's inputlist
is everything but practical.

This is an attempt to integrate fzf as a replacement for inputlist().
The main problem is that fzf doesn't run synchronously, at least I
didn't managed to get it to work like that!

So I had to rework then input_callback system of PHPActor vim plugin.
The idea is to use callable, that can be provided to fzf, to handle the
next step.
So if we have another input to submit to the user we call the input
dispatch another time and another one, until there is no more input.
Then we call the RPC with all the parameters.

Think of it as Middlewares if it helps, each input handle itself and
then call the next one.
The last one being the dispatcher to the RPC handler.