GenerateAccessorHandler can ask for a property name
Right know the accessor generator require the offset of the property to
generate an accessor for.
I think it's more practical to have the possibility to generate an
accessor from anywhere in a file juste by gaving phpactor it's name.

So I change the handler to make the offset optional, if not provided it
will return an action asking to choose a property in a list.
Thanks to fzf list we can also generate multiple accessors at once.

If a property name is provided is offset is retrieved using
WorseReflection.