Adds multi selection list support
OverrideMethodHandler defines its ListInput as accepting a multi selection.
It has no impact for vim's inputlist() but fzf allow to select multiple
items.
The result will be one RPC call for each selected item.
Not the best for performance but phpactor is local so no network lag and
it allows to use existing handlers because they don't have to change.