New strategy to position the cursor after deletion
The previous solution was still problematic because if there is an
insertion after a deletion there is no way to know if this insertion is
a replacement, then the cursor should go back to it's original position,
or a totaly different text, then the cursor has no reason to move.

So I decided to stop thinking of the "deleted line position" and
reasoning about the relative position of the cursor after the deletion.
Working around lines that doesn't exist anymore is the buffer is too
complex so I will work with the lines that were there and didn't change.

After a deletion the cursor go the the next undeleted line.
So if we memorize the relative position of the cursor to this line
before the deletion, we can move it back at the end.