Update CodeTransformExtraExtension to register extra template paths
For each defined template path we will look for directories which name
is a number of 5 digit and consider this name as a PHP version ID.
For each of these directory which corresponds to a version under or
equal to the current PHP version ID we will add them as a loader to
Twig, ordered by ascendant version.

This way we will have something like:
<project_root>/.phpactor/templates
<xdg_config>/phpactor/templates
<app>/templates/70400
<app>/templates/70200
<app>/templates

Which represent the priority in which the templates will be looked for.

Easy to tests, I added `70400/Property.php.twig` in the `code-builder`
packages.
Override the `php.version_id` to 70200 and try to complete a constructor
parameters: the properies added will have a phpdoc block if they had a
type
Do the same thing with `php.verion_id` to 70400: the propery will have a
type and no phpdoc block