Move config init to plugin directory
If it's just about initializing global configuration we can (must ?) do
it in the `plugin` directory instead of the ftplugin
The idea is that a global configuraiton is not link to a buffer and
should always be accessible.

I also change the auto command to add a use statement on auto
completion.
I made it a buffer command so that it is not triggerd for all buffers
after opening one php file.
I put it into a group in order to be able to clean everything, just in
case because I'm not sure how vim behave if we change the filetype
manually multiple times.
I made it call a temporary function that will check if the configuration
allow to import a class automatically or not.
I did it like this because I didn't want to change the way the function
currently behave.
I also updated the condition to look for a buffer configuration.
It more in order to show how it works than anything else, since we start
to define commands, mappings and all that on a buffer level we should
also start to let the user the possibility to override our configuration
on a buffer level.