doc: remove @library so functions are private
We don't document our functions because we prefer to provide commands.
That's why none of our functions are currently documented.

This make sense since our plugin is not a library for other plugins.
So I removed the @library annotation in order to be able to document
functions, not to generate it but so as developer we can have a little
bit more information