bugfix: Fix behavior where located source not containing thing
unresolvable classes were not being suggsted for import in a Drupal
project which had a classmap/ - this is because the "finder", for
perforamnce reasons, only located the source for the given name. As
classmap class location is YOLO it always located the source and the
name was reported as "resolved".