fix(tests): ansi output causes test failures
On some terminals, e.g. gnome-terminal, kitty (ubuntu), some tests fail
because the output of process contain color codes. For example in
the following assertion:

    $this->assertStringContainsString('class ⟶Badger⟵', $process->getOutput());

The characters surrounding "Badger" will be in color and so the
assertion fails because there are color codes in the string.

This patch adds `--no-ansi` to the `-php-cs-fixer` test command so that
the output never has color codes.