fix tests (#2521)
* fix(tests): ansi output causes test failures

On some terminals, e.g. gnome-terminal, kitty (ubuntu), some tests fail
because the output of process contain color codes. For example in
the following assertion:

    $this->assertStringContainsString('class ⟶Badger⟵', $process->getOutput());

The characters surrounding "Badger" will be in color and so the
assertion fails because there are color codes in the string.

This patch adds `--no-ansi` to the `-php-cs-fixer` test command so that
the output never has color codes.

* fix(test): passing null to type string is deprecated

When error reporting includes deprecations, the follow notice is emitted
when running the tests:

```
Deprecated: SplFileInfo::__construct(): Passing null to parameter #1 ($filename) of type string is deprecated in vendor/phpspec/prophecy/src/Prophecy/Doubler/Generator/ClassCreator.php(44) : eval()'d code on line 6
```

The test has been updated to use Symfony's SplFileInfo class which
Prophecy has an internal patch to help avoid the deprecations.