Fix parsing/generating Windows-style file: URIs
This fixes features like "go to definition" on Windows.

In a file: URI, after the scheme, there are two slashes, then
the 'host' (usually empty), then another slash, then the 'path'.
For example:

  file:///foo/bar.php
  file:///C:/foo/bar.php

That's three leading slashes total, and the third slash is not
a part of the path, but a part of the URI format itself; it just
looks like that with Unix-style filesystem paths.

Unfortunately, this means you can't just prepend 'file://'
to a filesystem path and get a valid file: URI. Fortunately,
`TextDocumentUri::fromString(…)` accepts filesystem paths too,
so document that and use it in test cases. Replace invalid URIs
with two slashes in other test cases too to avoid confusion.

See: https://en.wikipedia.org/wiki/File_URI_scheme