gh-2722: Two fixes
gh-2722: Flush index on file save
---------------------------------

also remove dead class

gh-2723: Fix location functions/classes in temporary diagnostics file
---------------------------------------------------------------------

WR uses a WorkspaceLocator to locate not-yet-saved classes and functions
and this locator uses the WR "WorkspaceIndex" to scan the definitions
in these files.

This commit passes the stdin source code given to the diagnostics
process o this "index" ensuring that the latest class and function
definitions are taken into account in the diagnostics process.