Grouping with phpstan/doc-parser is not simple task
For now, withdraw from attempts to separate with newline changes between
different tags. The most naive approach to check for previous and
current tag, and render empty line if change is detected fails short,
when there are tags, we don't want to render – those will result in
excessive empty lines.

Signed-off-by: Bart Jaskulski <bjaskulski@protonmail.com>