gh-2913: avoid infinite loop when resolving constant type (#2943)
When we have a cycle such as `declare('Foo', Foo)` PHP will bail at
runtime because it will evaluate `Foo` before `declare`. Phpactor will
evalute `Foo` lazily - but even if it located it eagerly (bad for
performance) the indxer _will find_ the constant, because the indexer
doesn't know that the value causes a cycle.

for now we just avoid resolving values if they are QualifiedName's.